Preliminary Ntr rewrite.
Restructure Ntr command sending interface to use one slot for sending a
command. Add an enum for the name of the command for use in the
interface.

Add NtrUtility class for functions not part of the core Ntr debugging
functionality. Automatic parsing of pid list is handled in this class,
as well as functionality for NFC patching.

Create a thread in the main() function for the new Ntr class to run in.
Change other classes to accomodate the new interface.