Further restructuring of NTR debugger.
Improve signaling and slots.  Add state to Ntr and StreamWorker that are
signalled when connected and disconnected.  Try to reduce the total
number of connections.

Add separate button for debugging separate from streaming. Add Pokemon
Sun and Moon NFC patch.

Add condition on StreamWindow to close the window when scale is set to
0.